Release version 2.14.0 and update documentation
Bump version to 2.14.0 in configure.in and update NEWS with new features and bug fixes. Add autogen.sh to distribution files in Makefile.am and Makefile.in. Improve help string formatting in autoconf macros for ALSA, ESD, libao, libogg, and libvorbis. Update ChangeLog with recent changes and fixes. Minor formatting adjustments in COPYING.